requirements.txt
pyupbit~=0.2.34
pandas~=2.2.3
apscheduler~=3.10.4
psycopg2~=2.9.10
python-dotenv~=1.0.1
fastapi~=0.115.2
requests~=2.32.3